+++ generators/app/prompts/promptsDB.js
@@ -1,27 +1,48 @@
-import _ from 'lodash';
+'use strict';
 
-export const promptDBName = value => ({
-  type: 'input',
-  name: 'name',
-  message: 'Database Name:',
-  default: value,
-  filter: str => str.replace(' ', '-'),
-  validate: _.negate(_.isEmpty)
+Object.defineProperty(exports, "__esModule", {
+  value: true
 });
+exports.promptDBPassword = exports.promptDBUser = exports.promptDBName = undefined;
 
-export const promptDBUser = value => ({
-  type: 'input',
-  name: 'user',
-  message: 'Database User:',
-  default: value,
-  filter: str => str.replace(' ', '-'),
-  validate: _.negate(_.isEmpty)
-});
+var _lodash = require('lodash');
 
-export const promptDBPassword = value => ({
-  type: 'password',
-  name: 'password',
-  message: 'Database Password:',
-  filter: String,
-  default: value
-});
+var _lodash2 = _interopRequireDefault(_lodash);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var promptDBName = exports.promptDBName = function promptDBName(value) {
+  return {
+    type: 'input',
+    name: 'name',
+    message: 'Database Name:',
+    default: value,
+    filter: function filter(str) {
+      return str.replace(' ', '-');
+    },
+    validate: _lodash2.default.negate(_lodash2.default.isEmpty)
+  };
+};
+
+var promptDBUser = exports.promptDBUser = function promptDBUser(value) {
+  return {
+    type: 'input',
+    name: 'user',
+    message: 'Database User:',
+    default: value,
+    filter: function filter(str) {
+      return str.replace(' ', '-');
+    },
+    validate: _lodash2.default.negate(_lodash2.default.isEmpty)
+  };
+};
+
+var promptDBPassword = exports.promptDBPassword = function promptDBPassword(value) {
+  return {
+    type: 'password',
+    name: 'password',
+    message: 'Database Password:',
+    filter: String,
+    default: value
+  };
+};